+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added support for gRPC server side reflection to all components (#949).
 
+### Changes
+
+- Added a faucet Dockerfile (#933).
+
 ## v0.9.0 (2025-05-30)
 
 ### Enhancements